remove underscore in class name
