+++ UnitTest/DataStructuresTests/GraphsDirectedSparseGraphTest.cs
@@ -6,11 +6,11 @@ using DataStructures.Graphs;
 
 namespace C_Sharp_Algorithms.DataStructuresTests
 {
-    public static class Graphs_UndirectedSparseGraphTest
+    public static class GraphsDirectedSparseGraphTest
     {
         public static void DoTest()
         {
-            var graph = new UndirectedSparseGraph<string>();
+            var graph = new DirectedSparseGraph<string>();
 
             var verticesSet1 = new string[] { "a", "z", "s", "x", "d", "c", "f", "v" };
 
@@ -21,37 +21,41 @@ namespace C_Sharp_Algorithms.DataStructuresTests
             graph.AddEdge("s", "x");
             graph.AddEdge("x", "d");
             graph.AddEdge("x", "c");
+            graph.AddEdge("x", "a");
             graph.AddEdge("d", "f");
             graph.AddEdge("d", "c");
+            graph.AddEdge("d", "s");
             graph.AddEdge("c", "f");
             graph.AddEdge("c", "v");
+            graph.AddEdge("c", "d");
             graph.AddEdge("v", "f");
+            graph.AddEdge("f", "c");
 
             var allEdges = graph.Edges.ToList();
 
             Debug.Assert(graph.VerticesCount == 8, "Wrong vertices count.");
-            Debug.Assert(graph.EdgesCount == 10, "Wrong edges count.");
+            Debug.Assert(graph.EdgesCount == 14, "Wrong edges count.");
             Debug.Assert(graph.EdgesCount == allEdges.Count, "Wrong edges count.");
 
             Debug.Assert(graph.OutgoingEdges("a").ToList().Count == 2, "Wrong outgoing edges from 'a'.");
-            Debug.Assert(graph.OutgoingEdges("s").ToList().Count == 2, "Wrong outgoing edges from 's'.");
-            Debug.Assert(graph.OutgoingEdges("x").ToList().Count == 3, "Wrong outgoing edges from 'x'.");
+            Debug.Assert(graph.OutgoingEdges("s").ToList().Count == 1, "Wrong outgoing edges from 's'.");
             Debug.Assert(graph.OutgoingEdges("d").ToList().Count == 3, "Wrong outgoing edges from 'd'.");
-            Debug.Assert(graph.OutgoingEdges("c").ToList().Count == 4, "Wrong outgoing edges from 'c'.");
-            Debug.Assert(graph.OutgoingEdges("v").ToList().Count == 2, "Wrong outgoing edges from 'v'.");
-            Debug.Assert(graph.OutgoingEdges("f").ToList().Count == 3, "Wrong outgoing edges from 'f'.");
-            Debug.Assert(graph.OutgoingEdges("z").ToList().Count == 1, "Wrong outgoing edges from 'z'.");
+            Debug.Assert(graph.OutgoingEdges("x").ToList().Count == 3, "Wrong outgoing edges from 'x'.");
+            Debug.Assert(graph.OutgoingEdges("c").ToList().Count == 3, "Wrong outgoing edges from 'c'.");
+            Debug.Assert(graph.OutgoingEdges("v").ToList().Count == 1, "Wrong outgoing edges from 'v'.");
+            Debug.Assert(graph.OutgoingEdges("f").ToList().Count == 1, "Wrong outgoing edges from 'f'.");
+            Debug.Assert(graph.OutgoingEdges("z").ToList().Count == 0, "Wrong outgoing edges from 'z'.");
 
-            Debug.Assert(graph.IncomingEdges("a").ToList().Count == 2, "Wrong incoming edges from 'a'.");
+            Debug.Assert(graph.IncomingEdges("a").ToList().Count == 1, "Wrong incoming edges from 'a'.");
             Debug.Assert(graph.IncomingEdges("s").ToList().Count == 2, "Wrong incoming edges from 's'.");
-            Debug.Assert(graph.IncomingEdges("x").ToList().Count == 3, "Wrong incoming edges from 'x'.");
-            Debug.Assert(graph.IncomingEdges("d").ToList().Count == 3, "Wrong incoming edges from 'd'.");
-            Debug.Assert(graph.IncomingEdges("c").ToList().Count == 4, "Wrong incoming edges from 'c'.");
-            Debug.Assert(graph.IncomingEdges("v").ToList().Count == 2, "Wrong incoming edges from 'v'.");
+            Debug.Assert(graph.IncomingEdges("d").ToList().Count == 2, "Wrong incoming edges from 'd'.");
+            Debug.Assert(graph.IncomingEdges("x").ToList().Count == 1, "Wrong incoming edges from 'x'.");
+            Debug.Assert(graph.IncomingEdges("c").ToList().Count == 3, "Wrong incoming edges from 'c'.");
+            Debug.Assert(graph.IncomingEdges("v").ToList().Count == 1, "Wrong incoming edges from 'v'.");
             Debug.Assert(graph.IncomingEdges("f").ToList().Count == 3, "Wrong incoming edges from 'f'.");
             Debug.Assert(graph.IncomingEdges("z").ToList().Count == 1, "Wrong incoming edges from 'z'.");
 
-            Console.WriteLine("[*] Undirected Sparse Graph:");
+            Console.WriteLine("[*] Directed Sparse Graph:");
             Console.WriteLine("Graph nodes and edges:");
             Console.WriteLine(graph.ToReadable() + "\r\n");
 
@@ -59,14 +63,14 @@ namespace C_Sharp_Algorithms.DataStructuresTests
             graph.RemoveEdge("c", "v");
             graph.RemoveEdge("a", "z");
             Debug.Assert(graph.VerticesCount == 8, "Wrong vertices count.");
-            Debug.Assert(graph.EdgesCount == 7, "Wrong edges count.");
+            Debug.Assert(graph.EdgesCount == 11, "Wrong edges count.");
 
             Console.WriteLine("After removing edges (d-c), (c-v), (a-z):");
             Console.WriteLine(graph.ToReadable() + "\r\n");
 
             graph.RemoveVertex("x");
             Debug.Assert(graph.VerticesCount == 7, "Wrong vertices count.");
-            Debug.Assert(graph.EdgesCount == 4, "Wrong edges count.");
+            Debug.Assert(graph.EdgesCount == 7, "Wrong edges count.");
 
             Console.WriteLine("After removing node (x):");
             Console.WriteLine(graph.ToReadable() + "\r\n");
@@ -75,21 +79,42 @@ namespace C_Sharp_Algorithms.DataStructuresTests
             graph.AddEdge("s", "x");
             graph.AddEdge("x", "d");
             graph.AddEdge("x", "c");
+            graph.AddEdge("x", "a");
             graph.AddEdge("d", "c");
             graph.AddEdge("c", "v");
             graph.AddEdge("a", "z");
             Console.WriteLine("Re-added the deleted vertices and edges to the graph.");
             Console.WriteLine(graph.ToReadable() + "\r\n");
 
-            Console.WriteLine("Walk the graph using BFS:");
-            graph.BreadthFirstWalk("s");		// output: (s) (a) (x) (z) (d) (c) (f) (v)
+            // BFS from A
+            Console.WriteLine("Walk the graph using BFS from A:");
+            var bfsWalk = graph.BreadthFirstWalk("a");		// output: (s) (a) (x) (z) (d) (c) (f) (v)
+            foreach (var node in bfsWalk) Console.Write(String.Format("({0})", node));
             Console.WriteLine("\r\n");
+
+            // DFS from A
+            Console.WriteLine("Walk the graph using DFS from A:");
+            var dfsWalk = graph.DepthFirstWalk("a");		// output: (s) (a) (x) (z) (d) (c) (f) (v)
+            foreach (var node in dfsWalk) Console.Write(String.Format("({0})", node));
+            Console.WriteLine("\r\n");
+
+            // BFS from F
+            Console.WriteLine("Walk the graph using BFS from F:");
+            bfsWalk = graph.BreadthFirstWalk("f");		// output: (s) (a) (x) (z) (d) (c) (f) (v)
+            foreach (var node in bfsWalk) Console.Write(String.Format("({0})", node));
+            Console.WriteLine("\r\n");
+
+            // DFS from F
+            Console.WriteLine("Walk the graph using DFS from F:");
+            dfsWalk = graph.DepthFirstWalk("f");		// output: (s) (a) (x) (z) (d) (c) (f) (v)
+            foreach (var node in dfsWalk) Console.Write(String.Format("({0})", node));
+            Console.WriteLine("\r\n");
+
+            Console.ReadLine();
 
 
             /********************************************************************/
 
-
-            Console.WriteLine("***************************************************\r\n");
 
             graph.Clear();
             Console.WriteLine("Cleared the graph from all vertices and edges.\r\n");
@@ -110,13 +135,17 @@ namespace C_Sharp_Algorithms.DataStructuresTests
             Debug.Assert(graph.VerticesCount == 6, "Wrong vertices count.");
             Debug.Assert(graph.EdgesCount == 8, "Wrong edges count.");
 
-            Console.WriteLine("[*] NEW Undirected Sparse Graph:");
+            Console.WriteLine("[*] NEW Directed Sparse Graph:");
             Console.WriteLine("Graph nodes and edges:");
             Console.WriteLine(graph.ToReadable() + "\r\n");
 
             Console.WriteLine("Walk the graph using DFS:");
-            graph.DepthFirstWalk();		// output: (a) (b) (e) (d) (c) (f) 
-        }
-    }
-}
+            dfsWalk = graph.DepthFirstWalk();		// output: (a) (b) (e) (d) (c) (f) 
+            foreach (var node in dfsWalk) Console.Write(String.Format("({0})", node));
 
+            Console.ReadLine();
+        }
+
+    }
+
+}
